NER/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
lmstudio-python>=1.3.0
pydantic>=2.7.0